requirements.txt
aiohttp==3.11.16
discord.py==2.5.2
yt_dlp==2025.3.31
